moving the extraction of logger to tags/logrus
