+++ logging/logrus/examples_test.go
@@ -3,8 +3,8 @@ package http_logrus_test
 import (
 	"net/http"
 
-	"github.com/improbable-eng/go-httpwares/logging/logrus"
 	"github.com/improbable-eng/go-httpwares/tags"
+	"github.com/improbable-eng/go-httpwares/tags/logrus"
 )
 
 var handler http.HandlerFunc
@@ -15,6 +15,6 @@ func ExampleExtract_withCustomTags() {
 		// Handlers can add extra tags to `http_ctxtags` that will be set in both the extracted loggers *and*
 		// the final log statement.
 		http_ctxtags.ExtractInbound(req).Set("my_custom.my_string", "something").Set("my_custom.my_int", 1337)
-		http_logrus.Extract(req).Warningf("Hello World")
+		ctx_logrus.Extract(req).Warningf("Hello World")
 	}
 }
