+++ logging/logrus/middleware.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 
 	"github.com/improbable-eng/go-httpwares"
+	"github.com/improbable-eng/go-httpwares/tags/logrus"
 	"github.com/sirupsen/logrus"
 )
 
@@ -19,18 +20,18 @@ var (
 
 // Middleware is a server-side http ware for logging using logrus.
 //
-// All handlers will have a Logrus logger in their context, which can be fetched using `http_logrus.Extract`.
+// All handlers will have a Logrus logger in their context, which can be fetched using `ctx_logrus.Extract`.
 func Middleware(entry *logrus.Entry, opts ...Option) httpwares.Middleware {
 	return func(nextHandler http.Handler) http.Handler {
 		o := evaluateMiddlewareOpts(opts)
 		return http.HandlerFunc(func(resp http.ResponseWriter, req *http.Request) {
 			wrappedResp := httpwares.WrapResponseWriter(resp)
 			newEntry := entry.WithFields(newServerRequestFields(req))
-			newReq := req.WithContext(toContext(req.Context(), newEntry))
+			newReq := req.WithContext(ctx_logrus.ToContext(req.Context(), newEntry))
 			var capture *responseCapture
 			wrappedResp.ObserveWriteHeader(func(w httpwares.WrappedResponseWriter, code int) {
 				if o.responseCaptureFunc(req, code) {
-					capture = captureMiddlewareResponseContent(w, Extract(newReq))
+					capture = captureMiddlewareResponseContent(w, ctx_logrus.Extract(newReq))
 				}
 			})
 			startTime := time.Now()
@@ -43,7 +44,7 @@ func Middleware(entry *logrus.Entry, opts ...Option) httpwares.Middleware {
 			}
 			level := o.levelFunc(wrappedResp.StatusCode())
 			levelLogf(
-				Extract(newReq).WithFields(postCallFields), // re-extract logger from newCtx, as it may have extra fields that changed in the holder.
+				ctx_logrus.Extract(newReq).WithFields(postCallFields), // re-extract logger from newCtx, as it may have extra fields that changed in the holder.
 				level,
 				"handled")
 		})
